+++ docs/changelog.rst
@@ -4,6 +4,13 @@ Changelog
 1.6   (unreleased)
 ------------------
 
+All deprecated code will be removed in the next release. Please run you project's tests using `python -Wd` so that
+deprecation warnings appear and can be addressed.
+
+This is last version of django-localflavor to support Django 1.8.x - 1.11.x which means this is the last version of
+django-localflavor that will support Python 2. Backwards-compatible bug fixes will be released as PATCH updates
+(e.g. 1.6.x) when needed.
+
 New flavors:
 
 - Added local flavor for Cuba